fix: restructure project layout
