--- .idea/sqldialects.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/dirMigrite/20241110191547-first.sql" dialect="GenericSQL" />
-    <file url="file://$PROJECT_DIR$/dirMigrite/20241110192953-first.sql" dialect="PostgreSQL" />
-    <file url="file://$PROJECT_DIR$/dirMigrite/20241110233543-first.sql" dialect="GenericSQL" />
-  </component>
-</project>
