+++ config.example.yaml
@@ -1,15 +1,12 @@
-
 app:
-  name: "MyApp"
-  version: "1.0.0"
+  name: "friends"
+  version: "0.0.1"
 
   development:
     server:
-
       httpPort: ":8080"
 
     database:
-
       address: "127.0.0.1:5432"
       username: "andrey"
       password: "andrey"
@@ -19,10 +16,3 @@ app:
       Datasource: "postgres://andrey:andrey@127.0.0.1:5432/database?sslmode=disable"
       dir: dirMigrite # Папка, где будут храниться файлы миграций
       table: migrations # Таблица для записи состояния миграций
-
-
-
-
-
-
-
