+++ internal/storage/dto.go
